Cache processed visualization data
