+++ frontend/src/context-object.tsx
@@ -1,6 +1,17 @@
-import { createContext } from "react";
+import React from "react";
 import Service from "./types/service.tsx";
 import { CookieKeys, Timescale } from "./types/strings";
+import GraphData from "./types/graph-data";
+import ChartData from "./types/chart-data";
+import ResourceUsageData from "./types/resource-usage-data";
+
+export interface ProcessedChartData {
+	quantityData: GraphData[];
+	responseCodeData: ChartData[];
+	countryCodeData: ChartData[];
+	IPAddressData: ChartData[];
+	resourceUsage: ResourceUsageData[];
+}
 
 export interface ContextType {
 	services: Service[];
@@ -9,11 +20,19 @@ export interface ContextType {
 	serviceAdd: (service: Service) => void;
 	serviceDelete: (serviceID: string) => void;
 	serviceUpdate: (service: Service) => void;
-	requestServiceData: () => void;
 	cookieGet: (key: CookieKeys) => string | undefined;
+	requestServiceData: () => void;
 	passwordReset: (newPassword: string) => void;
 	serviceToggle: (serviceID: string) => void;
+
+	// New chart data states for each timespan
+	hourData: ProcessedChartData;
+	dayData: ProcessedChartData;
+	monthData: ProcessedChartData;
+	yearData: ProcessedChartData;
+
+	// Function to get data for current timescale
+	getCurrentTimescaleData: () => ProcessedChartData;
 }
 
-// Create the context with a default value
-export const Context = createContext<ContextType | undefined>(undefined);
+export const Context = React.createContext<ContextType | undefined>(undefined);
